requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
boto3>=1.28.0
requests>=2.31.0
pydub>=0.25.1
pillow>=10.0.0
tqdm>=4.66.0
fastapi>=0.109.0
uvicorn>=0.27.0
